ファイルのアップロードで403になってた件を解決、他
wpコンテナでphp.iniを/usr/local/etc/php直下に置いていたことが原因っぽい
post_max_sizeとかを大き目に許容するための設定を書いてた
/usr/local/etc/php/conf.d以下にwordpress.iniという名前でコピーするよう修正したらファイルアップロードできた
↑できてなかった。wordpress.iniの設定はすべてコメントアウトすると正常にアップロードできた？？

wordpressのイメージで起動するとドキュメントルート下に日本語版でないwordpressが設置される
phpのイメージから起動するよう変更
wordpress:php7.4-fpm-alpine　→　php:7.4-fpm

+++ docker/wp/wp-install.sh
@@ -2,15 +2,19 @@
 
 set -ex;
 
-# WP_INSTALL_PATHがサブディレクトリの場合はホストPCと共有してるディレクトリを削除しようとしてエラーが出る。
-rm -fR /var/www/html/* || true
+# サブディレクトリにインストールする場合はここでwp-cliでWordPressをインストールする前にディレクトリ作成しておく
+if [ $WP_INSTALL_PATH != '/var/www/html' ]; then
+    # WP_INSTALL_PATHがサブディレクトリの場合はホストPCと共有してるディレクトリを削除しようとしてエラーが出る
+    # trueをパイプし、直前のコマンドの終了ステータスでDocker環境の起動が失敗するのを防ぐ
+    rm -fR /var/www/html/* || true
+    mkdir -p $WP_INSTALL_PATH
+fi
 
-mkdir -p $WP_INSTALL_PATH
 wp --allow-root core download --version=latest --locale=ja --path=$WP_INSTALL_PATH --force
 wp --allow-root core config --dbhost=db --dbname=$MYSQL_DATABASE --dbuser=$MYSQL_USER --dbpass=$MYSQL_PASSWORD --path=$WP_INSTALL_PATH
 wp --allow-root core install --title=$WP_BLOG_TITLE --admin_user=$WP_ADMIN_USER --admin_password=$WP_ADMIN_PASSWORD --admin_email=$WP_ADMIN_EMAIL --url=$WP_HTTP_ROOT --path=$WP_INSTALL_PATH
 wp --allow-root plugin uninstall akismet hello --path=$WP_INSTALL_PATH
 wp --allow-root plugin install wp-multibyte-patch --activate --path=$WP_INSTALL_PATH
+wp --allow-root plugin install classic-editor --activate --path=$WP_INSTALL_PATH
 wp --allow-root plugin install all-in-one-wp-migration --activate --path=$WP_INSTALL_PATH
-# 以下実行しても変わらない
-# chown -R www-data: $WP_INSTALL_PATH/wp-content
+
